Refactor and rename test_round_robin to remove RoundRobin class
